Create Pr0jectR3d-Music V 1.0.css
The 1.0. Release of the Pr0jectR3d-Music.
 
Here is the download link for easy installation: https://userstyles.world/style/8701/pr0jectr3d-music-v-0-1
This extension is required to apply the theme in a browser: https://chrome.google.com/webstore/detail/stylus/clngdbkpkpeebahjckkjfobafhncgmne

If you wish to use the Github variant, download the file and open it in a text editor Select the whole contents of the file. Copy it, and create a new project in the plugin listed above, and then paste the copied content in the project. Enjoy.